+++ test/libsolidity/syntaxTests/controlFlow/storageReturn/modifier_fine.sol
@@ -1,5 +1,5 @@
 contract C {
-    modifier revertIfNoReturn() {
+    modifier callAndRevert() {
         _;
         revert();
     }
@@ -9,12 +9,15 @@ contract C {
     }
     struct S { uint a; }
     S s;
-    function f(bool flag) revertIfNoReturn() internal view returns(S storage) {
+    function f(bool flag) callAndRevert() internal view returns(S storage) {
         if (flag) return s;
     }
-    function g(bool flag) revertIfNoReturn() ifFlag(flag) internal view returns(S storage) {
+    function g(bool flag) callAndRevert() ifFlag(flag) internal view returns(S storage) {
         return s;
     }
 
 }
 // ----
+// Warning 5740: (260-295): Unreachable code.
+// Warning 5740: (345-349): Unreachable code.
+// Warning 5740: (384-409): Unreachable code.